python_microservice/requirements.txt
uvicorn
fastapi
pydantic
requests
python-steamgriddb
